4 commit(Решил конфликт в Hibernate и БД изменил миграции, добавил Task)

+++ src/main/resources/db/migration/V1__init.sql
@@ -1,19 +1,20 @@
 CREATE TABLE users (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     email VARCHAR(255) UNIQUE NOT NULL,
     password VARCHAR(255) NOT NULL,
     role VARCHAR(50) NOT NULL
 );
 
 CREATE TABLE tasks (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     title VARCHAR(255) NOT NULL,
     description TEXT,
     status VARCHAR(50) NOT NULL,
     priority VARCHAR(50) NOT NULL,
-    author_id INT NOT NULL REFERENCES users(id),
-    executor_id INT REFERENCES users(id)
+    author_id BIGINT NOT NULL REFERENCES users(id),
+    executor_id BIGINT REFERENCES users(id)
 );
+
 
 CREATE TABLE comments (
     id SERIAL PRIMARY KEY,
